+++ packages/blade/src/components/Icons/Battery60PercentIcon/Battery60PercentIcon.native.test.tsx
@@ -1,0 +1,11 @@
+import Battery60PercentIcon from '.';
+import renderWithTheme from '~utils/testing/renderWithTheme.native';
+
+describe('<Battery60PercentIcon />', () => {
+  it('should render Battery60PercentIcon', () => {
+    const renderTree = renderWithTheme(
+      <Battery60PercentIcon color="feedback.icon.neutral.intense" size="large" />,
+    ).toJSON();
+    expect(renderTree).toMatchSnapshot();
+  });
+});