+++ packages/blade/src/components/StepGroup/types.ts
@@ -2,6 +2,7 @@ import type React from 'react';
 import type { BoxProps } from '~components/Box';
 import type { StyledPropsBlade } from '~components/Box/styledProps';
 import type { LinkProps } from '~components/Link';
+import type { FeedbackColors } from '~tokens/theme/theme';
 import type { DotNotationSpacingStringToken, TestID } from '~utils/types';
 
 type StepGroupProps = {
@@ -66,6 +67,11 @@ type StepItemProps = {
    * title of StepItem
    */
   title: string;
+
+  /**
+   * color of StepItem title
+   */
+  titleColor?: `feedback.text.${FeedbackColors}.intense` | `surface.text.primary.normal`;
 
   /**
    * A string that renders in italic font. Made for adding timestamp values.